+++ include/tbl/bo_flags.h
@@ -40,5 +40,6 @@ BO_FLAG(abandon,	0, 0, "")
 BO_FLAG(is_gzip,	0, 0, "")
 BO_FLAG(is_gunzip,	0, 0, "")
 BO_FLAG(was_304,	1, 0, "")
+BO_FLAG(is_bgfetch,	0, 0, "")
 
 /*lint -restore */